+++ .travis.yml
@@ -10,6 +10,7 @@ install:
   - pip install coveralls
 
 script:
+  - pytest -v --cov
   - coverage run -m pytest test/
 
 after_success: "coveralls"
