+++ project/__init__.py
@@ -23,8 +23,9 @@ migrate = Migrate(app, db)
 mail = Mail(app)
 
 from project.main.business import binnacle, company, eca_noise, gas_sensor, qhawax, eca_noise, qhawax_installation_history
-from project.main.data import processed_measurement, air_quality,gas_inca, valid_processed_measurement, air_daily_quality
+from project.main.data import processed_measurement, air_quality,gas_inca, valid_processed_measurement
+from project.main import util_helper
 import project.database.models as models
-from project.database.models import Company,GasSensor, Qhawax ,ProcessedMeasurement, AirQualityMeasurement, EcaNoise, GasInca, QhawaxInstallationHistory, ValidProcessedMeasurement,AirDailyMeasurement
+from project.database.models import Company,GasSensor, Qhawax ,ProcessedMeasurement, AirQualityMeasurement, EcaNoise, GasInca, QhawaxInstallationHistory, ValidProcessedMeasurement
 
 db.create_all()
